% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounding_box_for_iso_code.R
\name{crop_landcover_with_box}
\alias{crop_landcover_with_box}
\title{Crop a landcover raster with a bounding box}
\usage{
crop_landcover_with_box(landcover_raster, bounding_box, plot_invisible = FALSE)
}
\arguments{
\item{landcover_raster}{A \code{SpatRaster} object representing the landcover data to be cropped.}

\item{bounding_box}{A named list with four elements: \code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}, representing the coordinates of the bounding box to crop to.}

\item{plot_invisible}{Logical scalar indicating whether to plot the cropped raster.}
}
\value{
A \code{SpatRaster} object representing the cropped landcover data.
}
\description{
This function takes a landcover raster and a bounding box, and crops the raster to the extent of the bounding box. It also has an option to plot the cropped raster.
}
\examples{
crop_landcover_with_box(landcover_raster = landcover_trees, bounding_box = get_bounding_box_for_iso_code("NPL"), plot_invisible = TRUE)
}
