% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandUsePLN-package.R
\docType{package}
\name{LandUsePLN-package}
\alias{LandUsePLN}
\alias{LandUsePLN-package}
\title{LandUsePLN: Access Land Use Data in R}
\description{
This pipeline provides access to land use data endpoints. It is designed to be flexible and adaptable to various data sources and formats, allowing users to easily retrieve and analyze land use data for their specific needs. The current pipeline retrieves land use data from the [ESA World Cover](https://esa-worldcover.org/en) dataset.
}
\author{
\strong{Maintainer}: Tinashe Michael Tapera \email{ttapera@hsph.harvard.edu}

Other contributors:
\itemize{
  \item Christina Draeger \email{christina.draeger@posteo.de} [compiler]
  \item Christopher Golden \email{golden@hsph.harvard.edu} [copyright holder]
}

}
\keyword{internal}
